feat: 친밀도에 따른 이미지 변화 기능 구현

+++ src/main/java/capstone/facefriend/member/controller/FaceInfoController.java
@@ -28,7 +28,7 @@ public class FaceInfoController {
     @PutMapping("/face-info")
     public ResponseEntity<FaceInfoResponse> update(
             @RequestPart("origin")MultipartFile origin,
-            @RequestParam("styleId") Long styleId,
+            @RequestParam("styleId") Integer styleId,
             @AuthMember Long memberId
     ) throws IOException {
         return ResponseEntity.ok(faceInfoService.updateOrigin(origin, styleId, memberId));
